Various tidy ups
